requirements.txt
selenium
webdriver-manager
beautifulsoup4
pandas
mysql-connector-python
textblob
streamlit
